# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for TodoItem Application",
        "title": "TodoItem App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3333",
    "basePath": "/",
    "paths": {
        "/todo": {
            "get": {
                "description": "get latest TodoItem object",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoItem"
                ],
                "summary": "Get latest TodoItem",
                "operationId": "todo-latest",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.getLatestTodoItemResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.responseBag"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.responseBag"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.responseBag"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.getLatestTodoItemResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/todo.TodoItem"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.responseBag": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "todo.TodoItem": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        }
    }
}